# PingPong/PingPong.WebClient/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Pong",
    "version": "1.0"
  },
  "paths": {
    "/addMessage": {
      "post": {
        "tags": [
          "Message"
        ],
        "parameters": [
          {
            "name": "text",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageServiceResponse"
                }
              }
            }
          }
        }
      }
    },
    "/getMessage": {
      "get": {
        "tags": [
          "Message"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "messageId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageServiceResponse"
                }
              }
            }
          }
        }
      }
    },
    "/getMessageList": {
      "get": {
        "tags": [
          "Message"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "messageId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageListServiceResponse"
                }
              }
            }
          }
        }
      }
    },
    "/deleteMessage": {
      "delete": {
        "tags": [
          "Message"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "messageId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Int32ServiceResponse"
                }
              }
            }
          }
        }
      }
    },
    "/addUsers": {
      "post": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsersServiceResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Int32ServiceResponse": {
        "type": "object",
        "properties": {
          "response": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "Message": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "text": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "MessageListServiceResponse": {
        "type": "object",
        "properties": {
          "response": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Message"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "MessageServiceResponse": {
        "type": "object",
        "properties": {
          "response": {
            "$ref": "#/components/schemas/Message"
          }
        },
        "additionalProperties": false
      },
      "Users": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UsersServiceResponse": {
        "type": "object",
        "properties": {
          "response": {
            "$ref": "#/components/schemas/Users"
          }
        },
        "additionalProperties": false
      }
    }
  }
}